Update nutrition function

+++ datatypes.go
@@ -18,7 +18,7 @@ type SphereCell struct {
 	radius                           float64
 	position, velocity, acceleration OrderedPair
 	red, green, blue                 uint8
-	cellNutrition                    int
+	cellNutrition                    float64
 }
 
 // OrderedPair contains two float64 fields corresponding to
